Added a button to transition from the welcome screen to the interface
Sorry for the confusion! I was working on the code for a few hours
without realizing that I was working outside the cloned repository. The
changes I made are as follows:

- Made it so that the interface divs are originally set with a style of
display:none which means that they are invisible and do not affect the
layout of other objects
-Created a setWelcomeScreen function that does what it sounds like it
does
-Created a function called changeVisibility
-Created a button in the index that calls changeVisibility, which
subsequently makes the divs that were originally set as display:none to
display:block

Working on transitions and positioning